requirements.txt
streamlit>=1.28.0
openai>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
Pillow==12.0.0
camelot-py[base]>=0.11.0
langchain-text-splitters>=0.0.1
python-dateutil>=2.8.0


